backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
requests
gunicorn
